requirements.txt
# Core libraries
asgiref
python-dotenv
pillow
requests
streamlit
uvicorn[standard]
a2a-sdk
starlette
sse-starlette

# WhatsApp Integration
flask
twilio
pydub

# Google Cloud Speech-to-Text
google-cloud-speech

# The official Google Agent Development Kit (ADK)
google-adk
